+++ src/RoomDetail/types/type.ts
@@ -1,0 +1,3 @@
+export interface FormCertificationImage {
+  file: null | FileList;
+}
